fix list

+++ VolcomInventory/FormImportExcel.vb
@@ -1800,16 +1800,27 @@ Public Class FormImportExcel
             Dim tb1 = data_temp.AsEnumerable()
             Dim tb2 = dt.AsEnumerable()
 
+            'list
+            Dim ql As String = "SELECT dd.id_design, dm.design_code
+            FROM tb_emp_uni_design_det dd
+            INNER JOIN tb_m_design dm ON dm.id_design = dd.id_design
+            INNER JOIN tb_emp_uni_design d ON d.id_emp_uni_design = dd.id_emp_uni_design
+            WHERE d.id_emp_uni_period=" + FormEmpUniListDet.LEPeriodx.EditValue.ToString + "
+            AND d.id_report_status!=5 "
+            Dim dl As DataTable = execute_query(ql, -1, True, "", "", "", "")
+            Dim tb3 = dl.AsEnumerable()
+
             Dim query = From table1 In tb1
-                        Group Join table_tmp In tb2 On table1("KODE").ToString Equals table_tmp("design_code").ToString
-                            Into Group
-                        From y1 In Group.DefaultIfEmpty()
+                        Group Join table_tmp In tb2 On table1("KODE").ToString Equals table_tmp("design_code").ToString Into designjoin = Group
+                        From rd In designjoin.DefaultIfEmpty()
+                        Group Join list In tb3 On table1("KODE").ToString Equals list("design_code").ToString Into listjoin = Group
+                        From rl In listjoin.DefaultIfEmpty()
                         Select New With
                             {
-                                .IdDesign = If(y1 Is Nothing, "0", y1("id_design").ToString),
-                                .Code = If(y1 Is Nothing, table1("KODE"), y1("design_code").ToString),
-                                .Description = If(y1 Is Nothing, "-", y1("design_display_name").ToString),
-                                .Status = If(y1 Is Nothing, "Not Found", "OK")
+                                .IdDesign = If(rd Is Nothing, "0", rd("id_design").ToString),
+                                .Code = If(rd Is Nothing, table1("KODE"), rd("design_code").ToString),
+                                .Description = If(rd Is Nothing, "-", rd("design_display_name").ToString),
+                                .Status = If(rd Is Nothing, "Not Found", If(rl Is Nothing, "OK", "Already exist on this period"))
                             }
 
             GCData.DataSource = Nothing
@@ -3173,26 +3184,33 @@ Public Class FormImportExcel
                 Dim confirm As DialogResult = DevExpress.XtraEditors.XtraMessageBox.Show("Please make sure :" + System.Environment.NewLine + "- Only 'OK' status will continue to next step." + System.Environment.NewLine + "- If this report is an important, please click 'No' button, and then click 'Print' button to export to multiple formats provided." + System.Environment.NewLine + "Are you sure you want to continue this process?", "Warning", MessageBoxButtons.YesNo, MessageBoxIcon.Warning, MessageBoxDefaultButton.Button2)
                 Dim id_emp_uni_design As String = FormEmpUniListDet.id_emp_uni_design
                 If confirm = Windows.Forms.DialogResult.Yes Then
-                    'delete old
-                    Dim qd As String = "DELETE FROM tb_emp_uni_design_det WHERE id_emp_uni_design='" + id_emp_uni_design + "' "
-                    execute_non_query(qd, True, "", "", "", "")
+                    makeSafeGV(GVData)
+                    GVData.ActiveFilterString = "[Status] = 'OK'"
+                    If GVData.RowCount > 0 Then
+                        'delete old
+                        Dim qd As String = "DELETE FROM tb_emp_uni_design_det WHERE id_emp_uni_design='" + id_emp_uni_design + "' "
+                        execute_non_query(qd, True, "", "", "", "")
 
-                    PBC.Properties.Minimum = 0
-                    PBC.Properties.Maximum = GVData.RowCount - 1
-                    PBC.Properties.Step = 1
-                    PBC.Properties.PercentView = True
-                    '
-
-                    For i As Integer = 0 To ((GVData.RowCount - 1) - GetGroupRowCount(GVData))
-                        Dim id_design As String = GVData.GetRowCellValue(i, "IdDesign").ToString
-                        Dim query As String = "INSERT INTO tb_emp_uni_design_det(id_emp_uni_design, id_design) VALUES('" + id_emp_uni_design + "', '" + id_design + "'); "
-                        execute_non_query(query, True, "", "", "", "")
+                        PBC.Properties.Minimum = 0
+                        PBC.Properties.Maximum = GVData.RowCount - 1
+                        PBC.Properties.Step = 1
+                        PBC.Properties.PercentView = True
                         '
-                        PBC.PerformStep()
-                        PBC.Update()
-                    Next
-                    FormEmpUniListDet.viewDetail()
-                    Close()
+
+                        For i As Integer = 0 To ((GVData.RowCount - 1) - GetGroupRowCount(GVData))
+                            Dim id_design As String = GVData.GetRowCellValue(i, "IdDesign").ToString
+                            Dim query As String = "INSERT INTO tb_emp_uni_design_det(id_emp_uni_design, id_design) VALUES('" + id_emp_uni_design + "', '" + id_design + "'); "
+                            execute_non_query(query, True, "", "", "", "")
+                            '
+                            PBC.PerformStep()
+                            PBC.Update()
+                        Next
+                        FormEmpUniListDet.viewDetail()
+                        Close()
+                    Else
+                        stopCustom("There is no data for import process, please make sure your input !")
+                        makeSafeGV(GVData)
+                    End If
                 End If
             End If
         End If
